backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
gTTS==2.5.1
pydub==0.25.1
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
python-dotenv==1.0.1
imageio-ffmpeg
